refactor(lsp): split the 1.8k-line lsp_server.rs into focused modules
Pure code motion into latexml_oxide/src/lsp_server/ — public path
(lsp_server::run_lsp_server) unchanged:

- mod.rs         (61)  architecture overview, mod decls, run_lsp_server
- json.rs        (77)  serde_json Value shims + round-trip tests
- diagnostics.rs (167) Severity/Diag, log->diagnostic parsing + tests
- protocol.rs    (233) result/response/notification builders, framing
                       send, conversion-trigger extraction + coalescing
- document.rs    (265) uri->path, comment-aware preamble split,
                       sources, post-process, Config, dep snapshot
- server.rs      (102) warm-cache state + in-process fallback
- unix.rs        (~870) warm-fork pipeline, poll loop, dispatch
- generic.rs     (124) non-unix blocking fallback server

Internal items are pub(crate), re-exported via mod.rs globs so the
submodules keep the old single-module ergonomics (`use super::*`).
Tests moved next to their subjects (json/diagnostics/document/protocol;
framing_tests stay in unix.rs).

Two non-motion deltas, both review fallout:
- The fork-safety debug assertion from the previous commit asserted ONE
  live thread — but the latexml_oxide binary always runs the server on
  a dedicated worker thread while `main` parks in join() (2 threads,
  fork-benign), so it panicked on the first real convert. It now
  snapshots BASELINE_THREADS at server start and asserts the count
  never GROWS by fork time (the actual invariant).
- 00_unit_state.rs: bool-literal asserts -> assert!/assert!(!) (clippy).

Verified: 13/13 lsp unit tests; extended live smoke green end-to-end —
initialize, convert-preempts-convert (id2 "cancelled", id3 real HTML
for the newer text), malformed convert answered -32602, clean exit.
